requirements.txt
jira-python==3.5.2
python-dotenv==1.0.1 